### 2020-03-05 #2144 @csoneson
👋 @MetabolomicsAustralia-Bioinformatics - I'll handle your submission, and start by looking for suitable reviewers.

Looking at your submission, I noticed that your paper is a bit too short for JOSS standards (see https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain). Could you add a bit more context?
