### 2020-03-10 #2144 @csoneson
ping @bovee, @lgatto, @jorainer

> 👋 @bovee, @lgatto, @jorainer - would (two of) you be willing to review this submission for the Journal of Open Source Software (JOSS)?
>
> > MASTR-Quant: An Open-Source Web-based Software Tool for the Quantitive Analysis of Mass Spectrometry-based Data
>
> software: https://github.com/MetabolomicsAustralia-Bioinformatics/metabio
> accompanying paper: https://github.com/openjournals/joss-papers/blob/joss.02144/joss.02144/10.21105.joss.02144.pdf
>
> instructions on reviewing for JOSS: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
