### 2020-03-12 #2144 @csoneson
@sgibb - would you be interested in reviewing this submission for the Journal of Open Source Software?

> MASTR-Quant: An Open-Source Web-based Software Tool for the Quantitive Analysis of Mass Spectrometry-based Data

software: https://github.com/MetabolomicsAustralia-Bioinformatics/metabio
accompanying paper: https://github.com/openjournals/joss-papers/blob/joss.02144/joss.02144/10.21105.joss.02144.pdf

instructions on reviewing for JOSS: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
