### 2020-03-16 #2163 @csoneson
@lgatto, @bovee - thanks for accepting to review this submission! This is where the actual review takes place - your checklists are in the first post above, together with some more instructions. Just ping me if you have any questions.

@lgatto - I'll paste your comments from the PRE-REVIEW issue below.

@MetabolomicsAustralia-Bioinformatics - please have a look at these comments. As I also mentioned in the PRE-REVIEW issue, the paper is currently a bit short for JOSS standards, and needs a bit more context.

And of course, as Arfon already commented in the PRE-REVIEW issue, due to the COVID-19 pandemic JOSS currently handles existing manuscripts (such as this one) on a "best efforts basis". We understand that you may need to attend to more pressing issues than completing a review or updating a repository in response to a review. If this is the case, a quick note indicating that you need to put a "pause" on your involvement with a review would be appreciated but is not required.
