### 2020-03-16 #2163 @csoneson
Comments from @lgatto, transferred from the PRE-REVIEW issue:

Here are my comments regarding this submission:

* The submission contains a repository (but see below) and a valid OSI approved license.
* I find the manuscript to be extremely short. In particular, I miss some citation/comparison with existing alternatives in general, and a description of this software compares to other commonly-used packages in this research area.
* There are no references.
* The documentation on the README page only contains installation instructions (which I have not tested). There is no information about the underlying implementations or core functionality. In particular, it seems the user is let to decide what regression type and parameters are best suited based on visual inspection of the fit, being also allowed to manually define outliers. More rigorous approaches such (standardised) residuals, leverage and/or Cook's distance might be applicable.
* The application is available at mastr-quant.bio21.unimelb.edu.au, which contains some basic documentation of how to use the web interface. I tried it twice but the application froze and became unresponsive. The web application contains test data for user to try the interface out, but there is no description of these test data.
* I managed to download result files, which seem to contain various forms of the data (raw, normalised, ...), at time poorly formatted (multiple table per sheet) and I could find any documentation describing the output file or how to use the results for downstream analyses.
* It is confusion that the repo to be reviewed is itself a fork of uncle-leng/metabio. There are only two contributors to either of the fork or original repository. It would be useful to document the contributions of the 9 authors.

I confirm that I have read the JOSS conflict of interest policy and that I have no COIs with reviewing this work.
