### 2020-03-17 #2163 @bovee
I've done a preliminary review of the software and have some suggestions for improvement, but it appears GitHub issues are disabled on the repo (and there's no other documentation about how to make suggestions) so I'm unable to proceed.
