### 2020-03-17 #2163 @MetabolomicsAustralia-Bioinformatics
Hi Roderick,

Thanks for your input. I’ve just enabled GitHub issues on the repo (a bit surprised it’s not enabled by default!)

Regards,

don
