### 2020-03-20 #2144 @MetabolomicsAustralia-Bioinformatics
Hi all,

Thanks for your time and effort in reviewing so far. Only about manuscript submission:

1. We actually have a larger, 2 page, manuscript written, as a pdf. I’ve restructured (in particular, cut down) this larger paper to the approximate size and formatting as recommended here: https://joss.theoj.org/papers/10.21105/joss.00388, where it looked like only a summary was needed. Should I be submitting the full manuscript instead?
2. As for lack of references: I’ve actually included references in paper.bib (https://github.com/MetabolomicsAustralia-Bioinformatics/metabio/blob/master/paper/paper.bib), as per the submission instructions.  Unfortunately, this won’t actually compile. In fact, the Whedon paper preview service provided won’t compile its own example either to contain References (the gala paper). An attempt as of a few minutes ago has received a 404 error.

Just to address the manuscript issues, any advice to get around these technical issues would be greatly appreciated.

Many thanks and regards,
Don Teng
Bioinformatician
Metabolomics Australia
University of Melbourne, Parkville, Melbourne, Victoria
