### 2020-03-20 #2163 @csoneson
> Hi all,
> Thanks for your time and effort in reviewing so far. Only about manuscript submission:
> 1. We actually have a larger, 2 page, manuscript written, as a pdf. I’ve restructured (in particular, cut down) this larger paper to the approximate size and formatting as recommended here: https://joss.theoj.org/papers/10.21105/joss.00388, where it looked like only a summary was needed. Should I be submitting the full manuscript instead?
> 2. As for lack of references: I’ve actually included references in paper.bib (https://github.com/MetabolomicsAustralia-Bioinformatics/metabio/blob/master/paper/paper.bib), as per the submission instructions.  Unfortunately, this won’t actually compile. In fact, the Whedon paper preview service provided won’t compile its own example either to contain References (the gala paper). An attempt as of a few minutes ago has received a 404 error.

> Just to address the manuscript issues, any advice to get around these technical issues would be greatly appreciated.

Hi @MetabolomicsAustralia-Bioinformatics (please post any further comments in this issue, the pre-review issue is closed).
A typical JOSS paper is around 250-1000 words (your summary is currently ~170). However, it is not just a matter of the length in itself. I think your introduction needs to be expanded a bit to be understandable by a diverse, non-specialist audience and to provide the necessary context. As mentioned, we'd also like to have a brief discussion/comparison to other software addressing related needs. Please check https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain for more information.

Could you provide more details about the compilation issues? What is going wrong? I am able to compile the paper on the preview service just now. However, you should cite the references in the paper to make them show up in the bibliography.
