### 2020-03-22 MetabolomicsAustralia-Bioinformatics/metabio#1 @bovee
Installation
I had some trouble getting this up and running on Python 3.8. `venv` should probably not be checked into version control (I don't think it'll be cross-compatible from linux to mac, etc), but `pip install -r requirements.txt` doesn't work because the versions of several of the packages (e.g. numpy) are too old for my computer so I was only able to run using a mix of this and my local packages.

I think this worked, but running `python manage.py runserver` locally and trying to load the sample file gave me an alert that said "can not show result!!!".  This didn't seem to impact anything else on first glance, but changing metabolites after that also triggered this. The following was on the command line:
```
Internal Server Error: /show_result/
Traceback (most recent call last):
  File "/.../.miniconda/lib/python3.6/site-packages/django/core/handlers/exception.py", line 35, in inner
    response = get_response(request)
  File "/.../.miniconda/lib/python3.6/site-packages/django/core/handlers/base.py", line 128, in _get_response
    response = self.process_exception_by_middleware(e, request)
  File "/.../.miniconda/lib/python3.6/site-packages/django/core/handlers/base.py", line 126, in _get_response
    response = wrapped_callback(request, *callback_args, **callback_kwargs)
  File "/.../metabio/metabioapp/views.py", line 233, in show_result
    generate_output(input_dic, './result_xlsx.xlsx', options)
  File "/.../metabio/metabioapp/conc.py", line 285, in generate_output
    input_dic = is_normalisation(input_dic, is_method_dic)
  File "/.../metabio/metabioapp/conc.py", line 106, in is_normalisation
    if is_method_dic[each_col] == 'None':
KeyError: 'metabolite1'
[16/Mar/2020 02:46:44] "POST /show_result/ HTTP/1.1" 500 17726
```

I think either updating the packages (and possibly code?) to the most recent versions or adding a Dockerfile could go a ways to fix this. I realize the main use-case is as a server, but it makes it hard to develop on as-is.
