### 2020-03-22 MetabolomicsAustralia-Bioinformatics/metabio#3 @bovee
Tests don't work
Running `python manage.py test` found a test case, but it appears to be broken? It would be nice if there was a note on how to run tests in the README too.
