### 2020-03-22 MetabolomicsAustralia-Bioinformatics/metabio#4 @bovee
Documentation on output format
Some documentation on the subtables in the results spreadsheet and how to use them (or what that format feeds into?) would be useful.
