### 2020-03-22 MetabolomicsAustralia-Bioinformatics/metabio#5 @bovee
Contributions?
It would be nice if the README had a note about the best way to contribute (and also report issues with the software via issue, presumably).
