### 2020-03-22 #2163 @MetabolomicsAustralia-Bioinformatics
Hi @csoneson: thanks for your prompt reply.

1. I'll be uploading our manuscript in the required format. Thanks!

2. I've just tried the paper preview service; pasting [this link](https://github.com/adrn/gala/tree/master/paper) into the "Paper repository address". It seems to compile fine, but clicking the "click here to download preview" link leads to a "Not Found" page.
