+++ paper/paper.md
@@ -43,6 +43,30 @@ Targeted analysis in mass spectrometry uses a calibration curve to derive the co
 
 ``MASTR-Quant`` is an open-source, web application that allows users to calculate the concentration values of features in mass spectrometry data through visualizing and defining calibration curves based on a range of adjustable parameters. ``MASTR-Quant`` presents the user with options including and excluding data points to define linear or quadratic calibration ranges. The tool also includes options to assign individual internal standards to specific analytes by applying weighting factors, background subtraction and dilution ratios, normalization of the data based on internal standards and other external measurements, and the option for assessing the quality of data through the calculation of coefficient of variation values of each sample group. The final output is downloadable as an Excel file with multiple sheets containing detailed results of the individual steps involved in the calculation process.
 
+# Statement of Need
+
+``MASTR-Quant`` provides an open-access, transparent way of calculating calibration curves, which would otherwise be inaccessible due to the need to utilize proprietary vendor software, which are usually only compatible with their respective mass spectrometry instruments as well. The methods used by MASTR-Quant are compatible across all vendors.
+
+# 1. Introduction
+
+A multi-omics experiment (e.g. genomics, transcriptomics, proteomics, metabolomics etc.) can be defined as the integration of data across multiple omics layers (Joyce et al., 2006). This can lead to new insights into mechanisms around the genesis of pathogens as well as the creation of novel therapeutics (Civelek et al., 2014; Hasin et al., 2017). Quantitative omics studies allow for more comprehensive integration than non-quantitative studies i.e. reporting is completed in the same units across all omics, and the integration becomes mathematically more robust. The abundance of an analyte is measured and compared against a commercially available standard to obtain either a relative or absolute quantification value. There are several vendor specific software available which are able to extract peak areas and provide absolute concentration values; including Agilent Technologies MassHunter Quantitative Analysis(TM), Shimadzu’s LabSolutions(TM), ThermoFisher’s Xcalibur(TM), Water’s Empower(TM) amongst others. However, being commercial options, they are not freely available and cannot be used for data collected from instruments other than the vendor-specific instruments. In addition to this, calculations completed through spreadsheets can be time consuming and error-prone (http://blogs.nature.com/naturejobs/2017/02/27/escape-gene-name-mangling-with-escape-excel/). This led us to develop a freely available open-source tool, MASTR-Quant that uses processed data from any instrument to obtain quantification results. This freely available software allows users to import data from any analytical instruments including Gas Chromatography Mass Spectrometry (GC-MS) or Liquid Chromatography Mass Spectrometry (LC-MS).
+
+# 2. Methods and Features
+
+MASTR-Quant allows users to efficiently calculate the concentration values of features in mass spectrometry data through visualizing and defining calibration curves based on adjustable parameters.
+
+Data is input to MASTR-Quant in the form of a .csv file that is populated with peak abundances of analytes, which are represented in columns and samples to which they belong represented in rows. A template file is presented in the website for users to download and populate. Any internal standards that may be used are named with a preceding “IS_” tag. Upon uploading an input file, the list of analytes is displayed in a clickable table and all internal standards are displayed in a secondary table. Upon selecting an analyte, its calibration curve is displayed in the main plot with clickable data points. At this point, selecting an internal standard from the secondary table, normalises the data points to the selected internal standard and the resulting calibration curve is plotted. Users can either assign individual internal standards to different analytes or assign the same internal standard to all analytes using the “assign” and “assign all” buttons. Using these options, users are able to assign multiple internal standards to analytes and hence at the end of the workflow they are able to normalise the data based on multiple internal standards.
+
[…]
