### 2020-03-23 100c2ba
added link to docs, modified readme
