### 2020-03-23 #2163 @csoneson
@MetabolomicsAustralia-Bioinformatics - the JOSS paper preview works fine for me if I just put the link to the repository (not the paper/ folder): https://github.com/adrn/gala
