### 2020-03-23 #2163 @arfon
Yes, the form expects just the repository address. Whedon can find the paper within the repository.
