+++ requirements.txt
@@ -1,7 +1,8 @@
 Django==2.0
 mysqlclient==1.3.12
 numpy==1.13.3
-pandas==0.22.0
+#pandas==0.22.0
+pandas
 PyMySQL==0.7.11
 python-dateutil==2.6.1
 pytz==2017.3
