### 2020-03-25 #2163 @MetabolomicsAustralia-Bioinformatics
Hi all,

Thanks for your input so far.

I’ve compiled and pushed the long-form paper.pdf to the repo, but it still lacks references, though paper.md and paper.bib are both in the same /paper dir in the repo.

I’m sensing that this is a silly syntax error on my part, but I can’t tell what I’m missing?

Many thanks and regards,
Don
