### 2020-04-15 #2163 @MetabolomicsAustralia-Bioinformatics
Hi all, thanks for the reminder.

There’s been a full paper uploaded to the repo; I just haven’t provided any updates since I’m in the process of addressing reviewer comments re. better documentation, that I’ve done so far with sphinx (link in README).

However, the new, longer version still doesn’t compile with a complete references section. Is there some syntax error I’m missing? The .md and .bib files I used are in the repo.

Thanks,
don
