### 2020-04-15 #2163 @csoneson
Thanks for the update @MetabolomicsAustralia-Bioinformatics. Did you try this:

> You need to use markdown syntax for the citations in `paper.md`. Like this (from the example paper):
>
> ```
> # Citations
>
> Citations to entries in paper.bib should be in
> [rMarkdown](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html)
> format.
>
> If you want to cite a software repository URL (e.g. something on GitHub without a preferred
> citation) then you can do it with the example BibTeX entry below for @fidgit.
>
> For a quick reference, the following citation commands can be used:
> - `@author:2001`  ->  "Author et al. (2001)"
> - `[@author:2001]` -> "(Author et al., 2001)"
> - `[@author1:2001; @author2:2001]` -> "(Author1 et al., 2001; Author2 et al., 2002)"
> ```
