### 2020-04-17 #2163 @bovee
Apologies for the delay; I made issues on the repo of things that I'd like to see before I can accept (in addition to the other changes in documentation and the paper requested), but forgot to link out here:

 - [ ] https://github.com/MetabolomicsAustralia-Bioinformatics/metabio/issues/1
 - [ ] https://github.com/MetabolomicsAustralia-Bioinformatics/metabio/issues/3
 - [ ] https://github.com/MetabolomicsAustralia-Bioinformatics/metabio/issues/4
 - [ ] https://github.com/MetabolomicsAustralia-Bioinformatics/metabio/issues/5
