### 2020-06-01 #2163 @MetabolomicsAustralia-Bioinformatics
Hi all,

Apologies for the radio silence. I’ve been swamped so far as the only bioinformatician/ML analyst in my organisation.

I am working on a testing suite with Django’s built-in suite on a local build. I understand that we would preferably get to a point of automated testing. I’ve done my best to address the other points raised far re.:documentation, explanations for a non-technical audience.

Thanks and regards,
Don
