### 2020-06-17 #2163 @csoneson
👋  @MetabolomicsAustralia-Bioinformatics - just checking in to clarify where we are here. Do I interpret the message above correctly in that you plan to add the test suite as well before the reviewers should take a second look at the submission?
