### 2020-06-26 #2163 @MetabolomicsAustralia-Bioinformatics
Many apologies, I thought I’d replied to this, but turns out that my email client (Spark) didn’t send out the reply!

Yes, please wait while a test-suite is written out. Apologies as well for any inconvenience.

Regards,
Don
