### 2020-10-18 #2163 @csoneson
@MetabolomicsAustralia-Bioinformatics - do you think you will have the time to restart working on this submission within the next month or so? If not, in order to respect the reviewers' time and availability committment, I will propose to close (reject) this submission, and invite a resubmission at a later point in time.
