### 2020-10-23 #2163 @csoneson
@openjournals/joss-eics - we have agreed to close this submission. Could you help with that?
